Фикс #1. Несогласованные типы данных
Исправил несогласованность типов данных, везде установил int, так как в ТЗ не указан тип. Поправить до int быстрее и надёжнее, так как float в питоне такой же как и float в postgreSQL и нужно писать и отлаживать проверки крайний случаев. int в питоне ограничен оперативкой, а проверки на int уже есть.

+++ backend/database/queries/wallet.py
@@ -24,7 +24,7 @@ async def get_wallet(db: AsyncSession, wallet_uuid: UUID) -> Wallet:
 
 
 async def update_wallet_balance(
-    db: AsyncSession, wallet_uuid: UUID, operation_type: str, amount: float
+    db: AsyncSession, wallet_uuid: UUID, operation_type: str, amount: int
 ) -> Wallet:
     async with acquire_lock(db, Wallet, str(wallet_uuid), "uuid") as wallet:
         if operation_type == "WITHDRAW":
